pyproject.toml
[tool.poetry]
name = "ecommerce"
version = "0.1.0"
description = "Web application for ecommerce"
authors = ["rafiwts <rafiwts@gmail.com>"]
readme = "README.md"

[tool.black]
exclude = "migrations"

[tool.ruff]
exclude = ["migrations"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
tomli = "^2.0.1"
pre-commit = "^3.5.0"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.9"
pillow = "^10.1.0"
django-phonenumber-field = "^7.3.0"
phonenumbers = "^8.13.27"
django-extensions = "^3.2.3"
pytest-django = "^4.8.0"
freezegun = "^1.5.0"
django-file-resubmit = "^0.5.2"
django-haystack = "^3.3.0"
elasticsearch = "7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
